Replacement of scipy random
The random module is depreciated in the current version of Scipy. This ties users to old versions of the software and can cause compatibility issues.

I replaced all instances of scipy's random module in the "en" folder to use Numpy's random.default_rng() object while having equivalent behavior to previous versions of niacin.